rename files after the module they instrument
