+++ tests/versioned/aws-sdk-v3/client-dynamodb.tap.js
@@ -45,7 +45,7 @@ tap.test('DynamoDB', (t) => {
     helper.registerInstrumentation({
       moduleName: '@aws-sdk/client-dynamodb',
       type: 'datastore',
-      onResolved: require('../../../lib/v3-dynamo-ddb')
+      onResolved: require('../../../lib/v3-client-dynamodb')
     })
 
     const lib = require('@aws-sdk/client-dynamodb')
