+++ tests/versioned/aws-sdk-v3/package.json
@@ -157,7 +157,7 @@
         }
       },
       "files": [
-        "dynamo-ddb.tap.js"
+        "client-dynamodb.tap.js"
       ]
     }
   ],
